Many Majors Changes
var y now scales linearly, not exponentially which means the scaling of
the center is also smaller. The model isn't cut off at the bottom
anymore, and now represents a 3-layer model which illustrates Programs
-> Cohorts -> Competencies. Many smaller utility functions that are
accessed by buildRoot(), which creates the tree for the sunburst to use.
Font is smaller, etc